Conspiracy theories
Neuschwabenland has been the subject of several conspiracy theories for decades. Most of them assume that in the wake of the German expedition of 1938–39, a huge military base was built there, into which several high-ranking Nazis and strong troop units withdrew in 1945. The US and Great Britain have supposedly been trying to conquer the area secretly and in vain for decades, and are said to have used nuclear weapons in this context. Proponents believe that survival of the Nazi troops is possible because of hot springs that provide energy and warmth.